HIVE-1748 Statistics broken for tables with size in excess of Integer.MAX_VALUE (Paul Butler via namit)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hive/trunk@1027666 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/StatsTask.java
@@ -334,7 +334,7 @@ public class StatsTask extends Task<StatsWork> implements Serializable {
                     Long.parseLong(parameters.get(StatsSetupConst.ROW_COUNT)) :
                     0L;
           long sz = parameters.containsKey(StatsSetupConst.TOTAL_SIZE) ?
-                    Integer.parseInt(parameters.get(StatsSetupConst.TOTAL_SIZE)) :
+                    Long.parseLong(parameters.get(StatsSetupConst.TOTAL_SIZE)) :
                     0L;
           if (hasStats) {
             PartitionStatistics oldPartStats = new PartitionStatistics(nf, nr, sz);
